Prototype - todolist
I follow a tutorial on how to create a todolist from youtube "Made With Matt".  I added some annotations. I am still learning react. I will change the code to reflect the todo list design that we want, but I figure this is atleast something to show to Thomas. I do not take credit for the code.